requirements.txt
# requirements.txt

quart
google-generativeai
pinecone
python-telegram-bot
python-dotenv
flask[async]
hypercorn